src/requirements.txt
flask
zappa
PyGithub
requests
ipaddress
PyYAML
